requirements.txt
rich
fabric2
pyyaml
pytest
pytest-pep8
ruamel.yaml
numpy
beartype

